+++ src/sentry/dynamic_sampling/tasks/boost_low_volume_transactions/base.py
@@ -1,72 +1,32 @@
-import logging
-import math
-from datetime import timedelta
-from typing import Dict, Optional, Sequence, Tuple
-
-import sentry_sdk
-
 from sentry import options, quotas
 from sentry.dynamic_sampling.models.base import ModelType
 from sentry.dynamic_sampling.models.common import RebalancedItem, guarded_run
 from sentry.dynamic_sampling.models.factory import model_factory
-from sentry.dynamic_sampling.models.projects_rebalancing import ProjectsRebalancingInput
 from sentry.dynamic_sampling.models.transactions_rebalancing import TransactionsRebalancingInput
-from sentry.dynamic_sampling.prioritise_projects import fetch_projects_with_total_volumes
-from sentry.dynamic_sampling.prioritise_transactions import (
+from sentry.dynamic_sampling.rules.base import (
+    is_sliding_window_enabled,
+    is_sliding_window_org_enabled,
+)
+from sentry.dynamic_sampling.rules.helpers.prioritise_project import (
+    get_prioritise_by_project_sample_rate,
+)
+from sentry.dynamic_sampling.rules.helpers.prioritize_transactions import (
+    set_transactions_resampling_rates,
+)
+from sentry.dynamic_sampling.rules.helpers.sliding_window import get_sliding_window_sample_rate
+from sentry.dynamic_sampling.tasks.boost_low_volume_transactions.utils import (
     ProjectTransactions,
     fetch_project_transaction_totals,
     fetch_transactions_with_total_volumes,
     get_orgs_with_project_counts,
     transactions_zip,
 )
-from sentry.dynamic_sampling.recalibrate_transactions import (
-    OrganizationDataVolume,
-    fetch_org_volumes,
-)
-from sentry.dynamic_sampling.rules.base import (
-    is_sliding_window_enabled,
-    is_sliding_window_org_enabled,
-)
-from sentry.dynamic_sampling.rules.helpers.prioritise_project import (
-    generate_prioritise_by_project_cache_key,
-    get_prioritise_by_project_sample_rate,
-)
-from sentry.dynamic_sampling.rules.helpers.prioritize_transactions import (
-    set_transactions_resampling_rates,
-)
-from sentry.dynamic_sampling.rules.helpers.sliding_window import (
-    SLIDING_WINDOW_CALCULATION_ERROR,
-    extrapolate_monthly_volume,
-    generate_sliding_window_cache_key,
-    generate_sliding_window_org_cache_key,
-    get_sliding_window_org_sample_rate,
-    get_sliding_window_sample_rate,
-    get_sliding_window_size,
-    mark_sliding_window_executed,
-    mark_sliding_window_org_executed,
-)
-from sentry.dynamic_sampling.rules.utils import (
-    DecisionDropCount,
-    DecisionKeepCount,
-    OrganizationId,
-    ProjectId,
-    adjusted_factor,
-    generate_cache_key_rebalance_factor,
-    get_redis_client_for_ds,
-)
-from sentry.dynamic_sampling.sliding_window import (
-    fetch_orgs_with_total_root_transactions_count,
-    fetch_projects_with_total_root_transactions_count,
-)
-from sentry.dynamic_sampling.snuba_utils import (
-    get_active_orgs,
-    get_orgs_with_project_counts_without_modulo,
-)
+from sentry.dynamic_sampling.tasks.constants import CACHE_KEY_TTL
 from sentry.dynamic_sampling.tasks.logging import log_sample_rate_source
-from sentry.models import Organization, Project
+from sentry.dynamic_sampling.tasks.utils import dynamic_sampling_task
+from sentry.models import Organization
 from sentry.tasks.base import instrumented_task
 from sentry.tasks.relay import schedule_invalidate_project_config
-from sentry.utils import metrics
 
 
 @instrumented_task(
@@ -74,16 +34,11 @@ from sentry.utils import metrics
     queue="dynamicsampling",
     default_retry_delay=5,
     max_retries=5,
-    soft_time_limit=2 * 60 * 60,  # 2hours
+    soft_time_limit=2 * 60 * 60,
     time_limit=2 * 60 * 60 + 5,
 )
-def prioritise_transactions() -> None:
-    """
-    A task that retrieves all relative transaction counts from all projects in all orgs
-    and invokes a task for rebalancing transaction sampling rates within each project
-    """
-    metrics.incr("sentry.tasks.dynamic_sampling.prioritise_transactions.start", sample_rate=1.0)
-
+@dynamic_sampling_task()
+def boost_low_volume_transactions() -> None:
     num_big_trans = int(
         options.get("dynamic-sampling.prioritise_transactions.num_explicit_large_transactions")
     )
@@ -91,23 +46,22 @@ def prioritise_transactions() -> None:
         options.get("dynamic-sampling.prioritise_transactions.num_explicit_small_transactions")
     )
 
-    with metrics.timer("sentry.tasks.dynamic_sampling.prioritise_transactions", sample_rate=1.0):
-        for orgs in get_orgs_with_project_counts(MAX_ORGS_PER_QUERY, MAX_PROJECTS_PER_QUERY):
-            # get the low and high transactions
-            for project_transactions in transactions_zip(
-                fetch_project_transaction_totals(orgs),
-                fetch_transactions_with_total_volumes(
-                    orgs,
-                    large_transactions=True,
-                    max_transactions=num_big_trans,
-                ),
-                fetch_transactions_with_total_volumes(
-                    orgs,
-                    large_transactions=False,
-                    max_transactions=num_small_trans,
-                ),
-            ):
-                process_transaction_biases.delay(project_transactions)
+    for orgs in get_orgs_with_project_counts():
+        # get the low and high transactions
+        for project_transactions in transactions_zip(
+            fetch_project_transaction_totals(orgs),
+            fetch_transactions_with_total_volumes(
+                orgs,
+                large_transactions=True,
+                max_transactions=num_big_trans,
+            ),
+            fetch_transactions_with_total_volumes(
+                orgs,
+                large_transactions=False,
+                max_transactions=num_small_trans,
+            ),
+        ):
+            boost_low_volume_transactions_of_project.delay(project_transactions)
 
 
 @instrumented_task(
@@ -115,14 +69,11 @@ def prioritise_transactions() -> None:
     queue="dynamicsampling",
     default_retry_delay=5,
     max_retries=5,
-    soft_time_limit=25 * 60,  # 25 mins
+    soft_time_limit=25 * 60,
     time_limit=2 * 60 + 5,
 )
-def process_transaction_biases(project_transactions: ProjectTransactions) -> None:
-    """
-    A task that given a project relative transaction counts calculates rebalancing
-    sampling rates based on the overall desired project sampling rate.
-    """
+@dynamic_sampling_task()
+def boost_low_volume_transactions_of_project(project_transactions: ProjectTransactions) -> None:
     org_id = project_transactions["org_id"]
     project_id = project_transactions["project_id"]
     total_num_transactions = project_transactions.get("total_num_transactions")
