+++ src/sentry/dynamic_sampling/tasks/recalibrate_orgs/base.py
@@ -2,11 +2,10 @@ import logging
 from datetime import timedelta
 from typing import Dict
 
-from sentry.dynamic_sampling.recalibrate_transactions import fetch_org_volumes
-from sentry.dynamic_sampling.rules.tasks.recalibrate_orgs.utils import rebalance_org
-from sentry.dynamic_sampling.snuba_utils import get_active_orgs
+from sentry.dynamic_sampling.tasks.common import get_active_orgs
+from sentry.dynamic_sampling.tasks.recalibrate_orgs.utils import fetch_org_volumes, rebalance_org
+from sentry.dynamic_sampling.tasks.utils import dynamic_sampling_task
 from sentry.tasks.base import instrumented_task
-from sentry.utils import metrics
 
 logger = logging.getLogger(__name__)
 
@@ -19,20 +18,18 @@ logger = logging.getLogger(__name__)
     soft_time_limit=2 * 60 * 60,  # 2hours
     time_limit=2 * 60 * 60 + 5,
 )
+@dynamic_sampling_task()
 def recalibrate_orgs() -> None:
     query_interval = timedelta(minutes=5)
-    metrics.incr("sentry.tasks.dynamic_sampling.recalibrate_orgs.start", sample_rate=1.0)
 
     # use a dict instead of a list to easily pass it to the logger in the extra field
     errors: Dict[str, str] = {}
+    for orgs in get_active_orgs(1000, query_interval):
+        for org_volume in fetch_org_volumes(orgs, query_interval):
+            error = rebalance_org(org_volume)
+            if error and len(errors) < 100:
+                error_message = f"organisation:{org_volume.org_id} with {org_volume.total} transactions from which {org_volume.indexed} indexed, generated error:{error}"
+                errors[str(org_volume.org_id)] = error_message
 
-    with metrics.timer("sentry.tasks.dynamic_sampling.recalibrate_orgs", sample_rate=1.0):
-        for orgs in get_active_orgs(1000, query_interval):
-            for org_volume in fetch_org_volumes(orgs, query_interval):
-                error = rebalance_org(org_volume)
-                if error and len(errors) < 100:
-                    error_message = f"organisation:{org_volume.org_id} with {org_volume.total} transactions from which {org_volume.indexed} indexed, generated error:{error}"
-                    errors[str(org_volume.org_id)] = error_message
-
-        if errors:
-            logger.info("Dynamic sampling organization recalibration failed", extra=errors)
+    if errors:
+        logger.info("Dynamic sampling organization recalibration failed", extra=errors)
