+++ src/sentry/dynamic_sampling/tasks/recalibrate_orgs/utils.py
@@ -1,16 +1,55 @@
-from typing import Optional
+import logging
+import time
+from dataclasses import dataclass
+from datetime import datetime, timedelta
+from typing import Generator, List, Optional
+
+from snuba_sdk import (
+    Column,
+    Condition,
+    Direction,
+    Entity,
+    Function,
+    Granularity,
+    Op,
+    OrderBy,
+    Query,
+    Request,
+)
 
 from sentry import quotas
-from sentry.dynamic_sampling.recalibrate_transactions import OrganizationDataVolume
 from sentry.dynamic_sampling.rules.utils import (
     adjusted_factor,
     generate_cache_key_rebalance_factor,
     get_redis_client_for_ds,
 )
+from sentry.dynamic_sampling.tasks.constants import (
+    MAX_REBALANCE_FACTOR,
+    MAX_SECONDS,
+    MIN_REBALANCE_FACTOR,
+)
+from sentry.sentry_metrics import indexer
+from sentry.snuba.dataset import Dataset, EntityKey
+from sentry.snuba.metrics.naming_layer.mri import TransactionMRI
+from sentry.snuba.referrer import Referrer
+from sentry.utils.snuba import raw_snql_query
 
-# MIN and MAX rebalance factor ( make sure we don't go crazy when rebalancing)
-MIN_REBALANCE_FACTOR = 0.1
-MAX_REBALANCE_FACTOR = 10
+logger = logging.getLogger(__name__)
+
+
+@dataclass(frozen=True)
+class OrganizationDataVolume:
+    """
+    Represents the total and indexed number of transactions received by an organisation
+    (in a particular interval of time).
+    """
+
+    # organisation id
+    org_id: int
+    # total number of transactions
+    total: int
+    # number of transactions indexed ( i.e. kept)
+    indexed: int
 
 
 def rebalance_org(org_volume: OrganizationDataVolume) -> Optional[str]:
@@ -54,3 +93,123 @@ def rebalance_org(org_volume: OrganizationDataVolume) -> Optional[str]:
         redis_client.delete(factor_key)
 
     return None
+
+
+def get_active_orgs(max_orgs: int, time_interval: timedelta) -> Generator[List[int], None, None]:
+    """
+    Fetch organisations in batches.
+    A batch will return at max max_orgs elements
+    """
+    start_time = time.time()
+    metric_id = indexer.resolve_shared_org(str(TransactionMRI.COUNT_PER_ROOT_PROJECT.value))
+    offset = 0
+
+    while (time.time() - start_time) < MAX_SECONDS:
+        query = (
+            Query(
+                match=Entity(EntityKey.GenericOrgMetricsCounters.value),
+                select=[
+                    Function("uniq", [Column("project_id")], "num_projects"),
+                    Column("org_id"),
+                ],
+                groupby=[
+                    Column("org_id"),
+                ],
+                where=[
+                    Condition(Column("timestamp"), Op.GTE, datetime.utcnow() - time_interval),
+                    Condition(Column("timestamp"), Op.LT, datetime.utcnow()),
+                    Condition(Column("metric_id"), Op.EQ, metric_id),
+                ],
+                granularity=Granularity(60),
+                orderby=[
+                    OrderBy(Column("org_id"), Direction.ASC),
+                ],
+            )
+            .set_limit(max_orgs + 1)
+            .set_offset(offset)
+        )
+        request = Request(
+            dataset=Dataset.PerformanceMetrics.value, app_id="dynamic_sampling", query=query
+        )
+        data = raw_snql_query(
+            request,
+            referrer=Referrer.DYNAMIC_SAMPLING_COUNTERS_GET_ACTIVE_ORGS.value,
+        )["data"]
+        count = len(data)
+        more_results = count > max_orgs
+        offset += max_orgs
+        if more_results:
+            data = data[:-1]
+
+        ret_val = []
+
+        for row in data:
+            ret_val.append(row["org_id"])
+
+        yield ret_val
+
+        if not more_results:
+            return
+
+
+def fetch_org_volumes(
+    org_ids: List[int], query_interval: timedelta
+) -> List[OrganizationDataVolume]:
+    """
+    Returns the number of total and indexed transactions received by all organisations in the
+    specified interval.
+    """
+    transaction_string_id = indexer.resolve_shared_org("decision")
+    transaction_tag = f"tags_raw[{transaction_string_id}]"
+    metric_id = indexer.resolve_shared_org(str(TransactionMRI.COUNT_PER_ROOT_PROJECT.value))
+    where = [
+        Condition(Column("timestamp"), Op.GTE, datetime.utcnow() - query_interval),
+        Condition(Column("timestamp"), Op.LT, datetime.utcnow()),
+        Condition(Column("metric_id"), Op.EQ, metric_id),
+        Condition(Column("org_id"), Op.IN, org_ids),
+    ]
+
+    keep_count = Function(
+        "sumIf",
+        [
+            Column("value"),
+            Function(
+                "equals",
+                [Column(transaction_tag), "keep"],
+            ),
+        ],
+        alias="keep_count",
+    )
+
+    ret_val: List[OrganizationDataVolume] = []
+
+    query = Query(
+        match=Entity(EntityKey.GenericOrgMetricsCounters.value),
+        select=[
+            Function("sum", [Column("value")], "total_count"),
+            Column("org_id"),
+            keep_count,
+        ],
+        groupby=[Column("org_id")],
+        where=where,
+        granularity=Granularity(60),
+        orderby=[
+            OrderBy(Column("org_id"), Direction.ASC),
+        ],
+    )
+    request = Request(
+        dataset=Dataset.PerformanceMetrics.value, app_id="dynamic_sampling", query=query
+    )
+    data = raw_snql_query(
+        request,
+        referrer=Referrer.DYNAMIC_SAMPLING_COUNTERS_GET_ORG_TRANSACTION_VOLUMES.value,
+    )["data"]
+
+    for row in data:
+        ret_val.append(
+            OrganizationDataVolume(
+                org_id=row["org_id"], total=row["total_count"], indexed=row["keep_count"]
+            )
+        )
+
+    return ret_val
