+++ src/sentry/dynamic_sampling/tasks/sliding_window_org/utils.py
@@ -1,8 +1,7 @@
 import logging
 import time
-from collections import defaultdict
 from datetime import datetime, timedelta
-from typing import List, Mapping, Sequence, Tuple
+from typing import List, Mapping
 
 from snuba_sdk import (
     Column,
@@ -17,7 +16,12 @@ from snuba_sdk import (
     Request,
 )
 
-from sentry.dynamic_sampling.rules.utils import OrganizationId, ProjectId
+from sentry.dynamic_sampling.rules.helpers.sliding_window import (
+    generate_sliding_window_org_cache_key,
+)
+from sentry.dynamic_sampling.rules.utils import OrganizationId, get_redis_client_for_ds
+from sentry.dynamic_sampling.tasks.common import compute_guarded_sliding_window_sample_rate
+from sentry.dynamic_sampling.tasks.constants import CACHE_KEY_TTL, CHUNK_SIZE, MAX_SECONDS
 from sentry.sentry_metrics import indexer
 from sentry.snuba.dataset import Dataset, EntityKey
 from sentry.snuba.metrics.naming_layer.mri import TransactionMRI
@@ -25,81 +29,26 @@ from sentry.snuba.referrer import Referrer
 from sentry.utils.snuba import raw_snql_query
 
 logger = logging.getLogger(__name__)
-MAX_SECONDS = 60
-CHUNK_SIZE = 9998  # Snuba's limit is 10000, and we fetch CHUNK_SIZE+1
 
 
-def fetch_projects_with_total_root_transactions_count(
-    org_ids: List[int], window_size: int
-) -> Mapping[OrganizationId, Sequence[Tuple[ProjectId, int]]]:
+def adjust_base_sample_rate_of_org(org_id: int, total_root_count: int, window_size: int) -> None:
     """
-    Fetches tuples of (org_id, project_id) and the respective root transaction counts.
+    Adjusts the base sample rate per org by considering its volume and how it fits w.r.t. to the sampling tiers.
     """
-    query_interval = timedelta(hours=window_size)
-    granularity = Granularity(3600)
-
-    count_per_root_metric_id = indexer.resolve_shared_org(
-        str(TransactionMRI.COUNT_PER_ROOT_PROJECT.value)
+    sample_rate = compute_guarded_sliding_window_sample_rate(
+        org_id, None, total_root_count, window_size
     )
-    where = [
-        Condition(Column("timestamp"), Op.GTE, datetime.utcnow() - query_interval),
-        Condition(Column("timestamp"), Op.LT, datetime.utcnow()),
-        Condition(Column("metric_id"), Op.EQ, count_per_root_metric_id),
-        Condition(Column("org_id"), Op.IN, list(org_ids)),
-    ]
+    # If the sample rate is None, we don't want to store a value into Redis, but we prefer to keep the system
+    # with the old value.
+    if sample_rate is None:
+        return
 
-    start_time = time.time()
-    offset = 0
-    aggregated_projects = defaultdict(list)
-    while (time.time() - start_time) < MAX_SECONDS:
-        query = (
-            Query(
-                match=Entity(EntityKey.GenericOrgMetricsCounters.value),
-                select=[
-                    Function("sum", [Column("value")], "root_count_value"),
-                    Column("org_id"),
-                    Column("project_id"),
-                ],
-                where=where,
-                groupby=[Column("org_id"), Column("project_id")],
-                orderby=[
-                    OrderBy(Column("org_id"), Direction.ASC),
-                    OrderBy(Column("project_id"), Direction.ASC),
-                ],
-                granularity=granularity,
-            )
-            .set_limit(CHUNK_SIZE + 1)
-            .set_offset(offset)
-        )
-
-        request = Request(
-            dataset=Dataset.PerformanceMetrics.value, app_id="dynamic_sampling", query=query
-        )
-
-        data = raw_snql_query(
-            request,
-            referrer=Referrer.DYNAMIC_SAMPLING_DISTRIBUTION_FETCH_PROJECTS_WITH_COUNT_PER_ROOT.value,
-        )["data"]
-
-        count = len(data)
-        more_results = count > CHUNK_SIZE
-        offset += CHUNK_SIZE
-
-        if more_results:
-            data = data[:-1]
-
-        for row in data:
-            aggregated_projects[row["org_id"]].append((row["project_id"], row["root_count_value"]))
-
-        if not more_results:
-            break
-    else:
-        logger.error(
-            f"Fetching the transaction root count of multiple orgs took more than {MAX_SECONDS} seconds.",
-            extra={"offset": offset},
-        )
-
-    return aggregated_projects
+    redis_client = get_redis_client_for_ds()
+    with redis_client.pipeline(transaction=False) as pipeline:
+        cache_key = generate_sliding_window_org_cache_key(org_id=org_id)
+        pipeline.set(cache_key, sample_rate)
+        pipeline.pexpire(cache_key, CACHE_KEY_TTL)
+        pipeline.execute()
 
 
 def fetch_orgs_with_total_root_transactions_count(
